Better concrete syntax and language extension support:
Introduced the concept of unmanaged parse tables, which can be used by
editors to parse with an extended syntax. For example, the
Stratego-Sugar editor now specifies that it can be used for any
Stratego* parse table, which means that it can use a
StrategoWebDSL.tbl from somewhere in the workspace when that syntax
demanded by a .meta file but no "proper" editor exists for it.

svn path=/spoofax-imp/trunk/org.strategoxt.imp.feature/; revision=20426